Refactoring: Rename ClientType class
