--- src/main/java/com/honeypot/domain/notification/dto/ClientType.java
@@ -1,7 +1,0 @@
-package com.honeypot.domain.notification.dto;
-
-public enum ClientType {
-
-    IOS, AOS, WEB
-
-}
