+++ src/main/java/com/honeypot/domain/notification/dto/NotificationTokenDto.java
@@ -1,5 +1,6 @@
 package com.honeypot.domain.notification.dto;
 
+import com.honeypot.domain.notification.entity.enums.ClientType;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -9,7 +10,7 @@ import java.time.LocalDateTime;
 @Getter
 public class NotificationTokenDto {
 
-    private Long notificationTokenId;
+    private String notificationTokenId;
 
     private Long memberId;
