+++ src/main/java/com/honeypot/domain/notification/dto/NotificationTokenUploadRequest.java
@@ -1,5 +1,6 @@
 package com.honeypot.domain.notification.dto;
 
+import com.honeypot.domain.notification.entity.enums.ClientType;
 import lombok.*;
 
 import javax.validation.constraints.NotBlank;
@@ -7,10 +8,10 @@ import javax.validation.constraints.NotNull;
 
 @Getter
 @Builder
+@ToString
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode
-@ToString
 public class NotificationTokenUploadRequest {
 
     @NotBlank
