+++ src/main/java/com/honeypot/domain/notification/entity/enums/ClientType.java
@@ -1,0 +1,7 @@
+package com.honeypot.domain.notification.entity.enums;
+
+public enum ClientType {
+
+    IOS, AOS, WEB
+
+}
